// src/components/card-event-pattern/card-event/card-event.interface.ts
import { GlobalVariants } from '../../../interfaces/global-variants';

export interface CardEventProps {
  imageSrc: string;
  imageAlt: string;
  headline: string;
  description: string;
  date?: string;
  location?: string;
  linkLabel?: string;
  linkSrc: string;
  variant?: GlobalVariants;
  className?: string;
}